refactor work
